Roll forward of the https://github.com/tensorflow/tensorflow/commit/4e4943edc3d2409bffb5776f99b941987d6eda82.
Added a condition in tf_inspect, that we don't attempt to set default for keywords that are not in args.

This means that we are not in fact representing the function with ArgSpec. However the same is true for Python3 implementation of inspect.getfullargspec:

def f(*args, **kwargs):
  return args
p = functools.partial(f, 2, b=7)

inspect.getfullargspec(p)

"FullArgSpec(args=[], varargs='args', varkw='kwargs', defaults=None, kwonlyargs=[], kwonlydefaults=None, annotations={})"

Also added one test in load_test.py.

PiperOrigin-RevId: 243218910