Use tf_stack in tf_decorator.py.
The implementation of extract_stack in tf_stack is more efficient than the built-in one, because it does not stat source code files unnecessarily. I add a limit argument to allow the callsite to remain the same.

Here are some timings in a colab. Note that the real-world performance improvement will probably be larger, as more files will need to be accessed and they are likely to be on a slower filesystem.
BEFORE: 10000 loops, best of 3: 117 ?s per loop
AFTER: 100000 loops, best of 3: 4.43 ?s per loop
PiperOrigin-RevId: 243277147