Perform redzone checking on the device instead of doing it on the host.
Redzone checking imposes a large overhead on convolution autotuning,
especially if done on the host.
This CL instead performs checking on the device,
which often halves the convolution autotuning time.

The checking blob is encoded in PTX directly as it is portable
(can be later used in tensorflow) and does not have a fixed-size-shape
limitation HLO has (checking with HLO is also fast, but then the code would
have to be recompiled for every different size).

If there is a redzone mismatch, the check is repeated on host, in order to
get a more useful error message.

PiperOrigin-RevId: 245844571