Fix for the broken `--config=rocm` build.
Note: This ia a different PR from PRs #28189 and #28263 (same symptom, different cause and fix)

The `--config=rocm` build was broken by the following commit.

https://github.com/tensorflow/tensorflow/commit/481748b5c92d217dd1b10725ce259ba8c7086df0

The changes made by the above commit had one bug in the changes for the ROCm platform, whias was leading to the build failure. Fixing that bug to make the `--config=rocm` build working again.