Update implementation_selector to swallow the error during function load.
The error is potentially a warning since the optimizer could run several times, and it might try to optimize the functions that was updated by function optimizer in previous run, which will fail due to mismatch function signature.

Change the error to be a warning since user don't need to worry about this issue.

PiperOrigin-RevId: 247264286