Fix LossScaleOptimizer with multiple GPUs if grads_and_vars is not a list/tuple.
DistributionStrategy has special logic when handling lists/tuples, and we need to trigger that logic for LossScaleOptimizer to work properly.

This fixes LossScaleOptimizer when used with a multi-GPU Python 3 Keras model compiled with cloning=False. That scenario causes the return value of zip() to be passed in as grads_and_vars, and in Python 3, the return value of zip() is not a list or tuple.

PiperOrigin-RevId: 247454647